add InputOutput updates

+++ src/MEL/Daq/Quanser/QuanserAI.cpp
@@ -29,7 +29,7 @@ bool QuanserAI::update() {
 
 bool QuanserAI::update_channel(ChanNum channel_number) {
     t_error result;
-    result = hil_read_analog(daq_.handle_, &channel_number, static_cast<uint32>(1), &values_[channel_number]);
+    result = hil_read_analog(daq_.handle_, &channel_number, 1, &values_[channel_number]);
     if (result == 0)
         return true;
     else {
@@ -59,7 +59,7 @@ bool QuanserAI::set_range(ChanNum channel_number, Voltage min_value, Voltage max
     if (!Module::set_range(channel_number, min_value, max_value))
         return false;
     t_error result;
-    result = hil_set_analog_input_ranges(daq_.handle_, &channel_number, static_cast<uint32>(1), &min_values_[channel_number], &max_values_[channel_number]);
+    result = hil_set_analog_input_ranges(daq_.handle_, &channel_number, 1, &min_values_[channel_number], &max_values_[channel_number]);
     if (result == 0) {
         LOG(Verbose) << "Set " << get_name() << " channel number " << channel_number << " range to min=" << min_value << ", max=" << max_value;
         return true;
