+++ src/MEL/Daq/Quanser/QuanserAO.cpp
@@ -30,7 +30,7 @@ bool QuanserAO::update() {
 
 bool QuanserAO::update_channel(ChanNum channel_number) {
     t_error result;
-    result = hil_write_analog(daq_.handle_, &channel_number, static_cast<uint32>(1), &values_[channel_number]);
+    result = hil_write_analog(daq_.handle_, &channel_number, 1, &values_[channel_number]);
     if (result == 0)
         return true;
     else {
@@ -60,7 +60,7 @@ bool QuanserAO::set_range(ChanNum channel_number, Voltage min_value, Voltage max
     if (!Module::set_range(channel_number, min_value, max_value))
         return false;
     t_error result;
-    result = hil_set_analog_output_ranges(daq_.handle_, &channel_number, static_cast<uint32>(1), &min_values_[channel_number], &max_values_[channel_number]);
+    result = hil_set_analog_output_ranges(daq_.handle_, &channel_number, 1, &min_values_[channel_number], &max_values_[channel_number]);
     if (result == 0) {
         LOG(Verbose) << "Set " << get_name() << " channel number " << channel_number << " range to min=" << min_value << ", max=" << max_value;
         return true;
@@ -92,7 +92,7 @@ bool QuanserAO::set_expire_value(ChanNum channel_number, Voltage expire_value) {
     if (!Output::set_expire_value(channel_number, expire_value))
         return false;
     t_error result;
-    result = hil_watchdog_set_analog_expiration_state(daq_.handle_, &channel_number, static_cast<uint32>(1), &expire_values_[channel_number]);
+    result = hil_watchdog_set_analog_expiration_state(daq_.handle_, &channel_number, 1, &expire_values_[channel_number]);
     if (result == 0) {
         LOG(Verbose) << "Set " << get_name() << " channel number " << channel_number << " expire value to " << expire_value;
         return true;
